requirements_auth.txt
# Requirements for Authenticated LinkedIn Scraper
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3